Twitter OAuth optimizations
During the OAuth authentication, when you exchange an OAuth verifier for an access token (and access token secret), Twitter will respond with a number of default properties as described by the OAuth 1.0a specification, but also two custom properties - that is the "user_id" and "screen_name" of the authenticated user.

Normally an instance of "OAuthAccessToken" would be returned for the response body, but with the two custom properties, an instance of "TwitterOAuthAccessToken" is now returned instead.

Although the response when getting a request token currently doesn't have any custom properties, the response body is now also represented by an instance of "TwitterOAuthRequestToken" rather than just "OAuthRequestToken".